% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIfNeeded.R
\name{downloadIfNeeded}
\alias{downloadIfNeeded}
\title{Download files if missing or incomplete}
\usage{
downloadIfNeeded(links, destdir)
}
\arguments{
\item{links}{Character vector of file download URLs.}

\item{destdir}{Path to the directory where files should be saved.}
}
\value{
Invisibly returns \code{NULL}. Files are written to \code{destdir} as a side effect.
}
\description{
Checks a set of file download links and downloads each file into a destination
directory if it is not already present or if the existing file appears
incomplete. File completeness is assessed by comparing the local file size
with the remote file size obtained via an HTTP HEAD request.
}
\details{
For each URL in \code{links}, this function:
\itemize{
\item Determines the expected file name using \code{basename()}.
\item Checks whether the file already exists in \code{destdir}.
\item If it exists, compares its size to the remote file size obtained
from the \code{content-length} header of an HTTP HEAD request.
\item Downloads the file only if it is missing or appears incomplete.
}

For large files, consider increasing the download timeout, e.g.:
\code{options(timeout = 6000)}.
}
\examples{
\dontrun{
links <- c(
  paste0("https://os.zhdk.cloud.switch.ch/chelsav2/GLOBAL/climatologies/",
  "1981-2010/bio/CHELSA_bio1_1981-2010_V.2.1.tif"),
  paste0("https://os.zhdk.cloud.switch.ch/chelsav2/GLOBAL/climatologies/";
  "1981-2010/bio/CHELSA_bio12_1981-2010_V.2.1.tif")
)

downloadIfNeeded(links, destdir = tempdir())


links <- c(links,
  paste0("https://os.zhdk.cloud.switch.ch/chelsav2/GLOBAL/climatologies/",
  "1981-2010/bio/CHELSA_bio8_1981-2010_V.2.1.tif"))

downloadIfNeeded(links, destdir = tempdir())
}

}
