% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkbuild.R
\name{linkbuild}
\alias{linkbuild}
\title{Build file download links}
\usage{
linkbuild(
  vars = paste0("bio", 1:19),
  prefix =
    "https://os.zhdk.cloud.switch.ch/chelsav2/GLOBAL/climatologies/1981-2010/bio/CHELSA_",
  suffix = "_1981-2010_V.2.1.tif"
)
}
\arguments{
\item{vars}{Character vector of variable identifiers. Defaults to
\code{paste0("bio", 1:19)}.}

\item{prefix}{Character string. URL prefix pointing to the file
location. Currently defaults to
\code{"https://os.zhdk.cloud.switch.ch/chelsav2/GLOBAL/climatologies/1981-2010/bio/CHELSA_"}.}

\item{suffix}{Character string appended after each variable name.
Currently defaults to \code{"_1981-2010_V.2.1.tif"}.}
}
\value{
A character vector of full URLs.
}
\description{
Constructs full download URLs (currently by default, for CHELSA v2
bioclimatic variables from the 1981-2010 climatologies) by pasting together
a prefix, variable names, and a suffix.
}
\examples{
linkbuild()

linkbuild(c("bio1", "bio12", "scd"))

}
\author{
A. Marcia Barbosa
}
