% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{Re-grid raster layer(s) onto an equal-area square-cell grid}
\usage{
regrid(
  layers,
  grid,
  fun = "mean",
  densif = 0,
  exactextract = FALSE,
  verbosity = 1,
  ...
)
}
\arguments{
\item{layers}{\code{SpatRaster} (or an object that can be coerced to it)
containing the input layer(s) to be re-gridded.}

\item{grid}{(\code{SpatRaster},) \code{SpatVector} or \code{sf} (the latter will be coerced
to \code{SpatVector}) defining the target grid. Must consist of square equal-area
polygons (e.g. the EEA reference grid), otherwise results will be incorrect.}

\item{fun}{aggregation/summarizing function (default "mean") passed to
\code{terra::zonal()}, or to \code{exactextractr::exact_extract()} if
exactextract = TRUE (see below).}

\item{densif}{zero (default) or positive integer value indicating the
number of equal parts in which to divide each line segment of \code{grid}, in
which case \code{terra::densify} is used to avoid the path changing too much when
projecting (if \code{layers} have a different CRS). This can make a difference
when grid cells are large and near polar latitudes.}

\item{exactextract}{logical (default FALSE) specifying whether the
extraction of \code{layers} values to the polygon \code{grid} should be performed
with the 'exactextractr' package rather than the default 'terra' package. Can be
considerably faster, but requires package 'exactextractr' (>= 0.10.1) to be
installed.}

\item{verbosity}{integer indicating the amount of progress messages to
display. The default is 1, for an intermediate amount of messages. Currently
meaningful values are integers from 0 to 2.}

\item{...}{Additional arguments passed to \code{terra::extract()}.}
}
\value{
The function returns a SpatRaster of the input \code{layers} re-gridded
to match the input \code{grid}.
}
\description{
This function extracts or summarizes (using zonal statistics) the values of
a set of raster layers onto an equal-area square-cell vector polygon grid,
then rasterizes the values onto a layer with the same extent, resolution and
coordinate reference system as the input vector grid. It is useful for
converting lon-lat variable layers into equal-area pixel grids (matching e.g. #' the EEA reference grid).
}
\details{
This function:
\itemize{
\item (if densif > 0) densifies the input vector \code{grid} by introducing additional vertices
\item projects the (densified) vector \code{grid} to properly overlay the input raster \code{layers}
\item extracts/summarizes \code{layers} values onto the projected \code{grid}, using
the function specified in \code{fun}
\item makes a SpatRaster grid with the same CRS and dimensions (extent, cell size) as the input polygon \code{grid}
\item rasterizes each extracted layer onto this SpatRaster grid
}
}
\examples{
\dontrun{

# download equal-area vector grid:

eea_url <- paste0("https://sdi.eea.europa.eu/datashare/s/jokLrYcEBFiJRyo/",
"download?path=\%2FGPKG&files=EEA_50km_grid_v2024.gpkg")

eea50 <- terra::vect(eea_url)


# subset grid to smaller area (for quicker example):

grid <- terra::subset(eea50, terra::ext(2.6e6, 3e6, 1.7e6, 2.4e6))

terra::plot(grid)


# import raster layers:

links <- linkbuild(c("bio1", "scd"))  # get CHELSA climate links

dir.create("outputs/variables", recursive = TRUE)

options(timeout = 6000)  # allow longer download times

downloadif(links, destdir = "outputs/variables")

layers <- terra::rast(list.files("outputs/variables", full.names = TRUE))

terra::plot(layers)


# re-grid layers:

out <- regrid(layers = layers, grid = grid, fun = "mean",
na.rm = TRUE, touches = TRUE)

terra::plot(out[[1:4]])


terra::plot(out[[1]])

terra::plot(grid, lwd = 0.2, add = TRUE)


# re-grid faster if you have 'exactextractr' installed:

out2 <- regrid(layers = layers, grid = grid, fun = "mean",
exactextract = TRUE)


terra::plot(out2[[1]])

terra::plot(grid, lwd = 0.2, add = TRUE)
}

}
\author{
A. Marcia Barbosa
}
