% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{Re-grid raster layer(s) onto an equal-area square-cell grid}
\usage{
regrid(layers, grid, fun = "mean", ...)
}
\arguments{
\item{layers}{A SpatRaster containing the input layer(s) to be re-gridded.}

\item{grid}{A polygon SpatVector defining the target grid. Must consist
of square equal-area cells (e.g. the EEA reference grid), otherwise
unexpected results may arise.}

\item{fun}{Character or function. Aggregation function passed to
\code{terra::extract()} (e.g., \code{"mean"}, \code{"median"}, \code{"sum"}).}

\item{...}{Additional arguments passed to \code{terra::extract()}.}
}
\value{
A SpatRaster with the same number of layers as \code{layers},
re-gridded to the resolution and extent of the target grid.
}
\description{
This function extracts the values of a set of raster layers onto a regular
equal-area vector polygon grid, then rasterizes the values onto a layer with
the same extent, resolution and coordinate reference system as the input
vector grid.
}
\details{
This function:
\itemize{
\item Projects the input \code{grid} to overlay on the input \code{layers}.
\item Extracts \code{layers} values onto the projected grid using \code{fun}.
\item Builds a SpatRaster with the same dimensions as \code{grid}.
\item Rasterizes each extracted layer onto this SpatRaster.
}
}
\examples{
\dontrun{
eea50 <- terra::vect("https://sdi.eea.europa.eu/datashare/s/jokLrYcEBFiJRyo/download?path=\%2FGPKG&files=EEA_50km_grid_v2024.gpkg")

ctry <- geodata::gadm(country = "Portugal", level = 0, path = tempdir())
ctry_prj <- terra::project(ctry, eea50)
grid <- terra::subset(eea50, ctry_prj)  # smaller object for faster computing
layers <- geodata::worldclim_global(var = "bio", res = 5, path = tempdir())

out <- regrid(layers = layers, grid = grid, fun = "mean")
}

}
